+++ face-rec-avenger/src/face_rec_avenger/pipeline_registry.py
@@ -1,51 +1,21 @@
 """Project pipelines."""
 from typing import Dict
-
-from kedro.framework.project import find_pipelines
 from kedro.pipeline import Pipeline, node
 
-from face_rec_avenger.nodes import (
-    extract_features,
-    train_model,
-    recognize_faces,
+from face_rec_avenger.pipelines import (
+    recognition_pipeline as rcp
 )
 
 
-def create_pipeline(**kwargs):
-    return Pipeline(
-        [
-            node(
-                extract_features,
-                ["train_images", "parameters"],
-                "features",
-                name="extract_features",
-            ),
-            node(
-                train_model,
-                ["features", "parameters"],
-                "svm_model",
-                name="train_model",
-            ),
-            node(
-                recognize_faces,
-                ["test_images", "svm_model", "features"],
-                "face_predictions",
-                name="recognize_faces",
-            ),
-        ]
-    )
-    
-    
-# def register_pipelines() -> Dict[str, Pipeline]:
-#     """Register the project's pipelines.
+def register_pipelines() -> Dict[str, Pipeline]:
+    """Register the project's pipelines.
 
-#     Returns:
-#         A mapping from pipeline names to ``Pipeline`` objects.
-#     """
-#     pipelines = find_pipelines()
-#     pipelines["__default__"] = sum(pipelines.values())
-#     return pipelines
+    Returns:
+        A mapping from pipeline names to ``Pipeline`` objects.
+    """
+    recognition_pipeline = rcp.create_pipeline()
 
-
-def register_pipelines():
-    return {"recognition_pipeline": create_pipeline()}
+    return {
+        "__default__": recognition_pipeline,
+        "recognition_pipeline": recognition_pipeline,
+    }
