+++ face-rec-avenger/src/face_rec_avenger/pipelines/recognition_pipeline.py
@@ -1,8 +1,8 @@
 from kedro.pipeline import Pipeline, node
-from .nodes import (
+from face_rec_avenger.nodes import (
     extract_features,
     classify,
-    recognize_faces,
+    predict,
 )
 
 
@@ -10,20 +10,20 @@ def create_pipeline(**kwargs):
     return Pipeline(
         [
             node(
-                extract_features,
-                ["train_images", "parameters"],
+                extract_features.extract_features,
+                ["test_images", "parameters"],
                 "features",
                 name="extract_features",
             ),
+            # node(
+            #     classify.classify,
+            #     ["features", "parameters"],
+            #     "svm_model",
+            #     name="train_model",
+            # ),
             node(
-                classify,
-                ["features", "parameters"],
-                "svm_model",
-                name="train_model",
-            ),
-            node(
-                recognize_faces,
-                ["test_images", "svm_model", "features"],
+                predict.predict,
+                ["svm_model", "features"],
                 "face_predictions",
                 name="recognize_faces",
             ),
